Add certmasteries to invtypes. Add sqlmock tests
Insert certmasteries when going over invtypes. This is done in a
separate transaction currently due to weirdnes when running 2 `COPY
FROM` in the same transaction.

Add testing of the insertions with sqlmock. These can't really be run in
the same order since iterating over map is not guaranteed to be
in-order.