Universe import at parity (almost)
Universe import at parity with the python importer, save for
`orbitIndex`.

Remove SQL calls for getItemNameByID and getGroupIDByTypeID, instead
look up straight in maps reflected from invNames and typeIDs (this means
that typeIDs must be imported before universe)

Denormalize data checks with the python importer except in the case of
one wormhole planet who's y axis is `-0` in the yaml but `0` in the
python import. Yay.